Reference

Reference

Signal Generation

SignalOperators.signal — Function.

signal(x,[samplerate])

Coerce x to be a signal, optionally specifying its sample rate (usually in Hz). Signal operations first coerce their arguments to be a signal so this needs only to be specified when the additional arguments to signal are needed.

Note

If you pipe signal (e.g. myobject |> signal(2kHz)) you must specify the units of the sample rate. This is because a raw number is ambiguous, and could be interpreted as a signal (i.e. an infinite length signal of with constant valued samples).

The types of objects that can be coerced to signals are as follows.

source

Existing signals

Any existing signal just returns itself from signal. If a sample rate is specified it will be set if x has an unknown sample rate. If it has a known sample rate and doesn't match samplerate(x) and error will be throwns. If you want to change the sample rate of a signal use tosamplerate.

source

Numbers

Numbers can be treated as infinite length, constant signals of unknown sample rate.

source

Arrays

Arrays can be treated as signals. The first dimension is time, the second channels.

AxisArrays, if they have an axis labled time and one or zero additioanl axes, can be treated as a signal. The time dimension must be represted using a Range.

source

Functions

signal(fn,[samplerate];[ω/frequency],[ϕ/phase])

Functions can define infinite length signals of known or unknown sample rate. The function fn can either return a number. Or, if you wish to specify a multi-channel signal, a tuple of values.

The input to fn is either a phase value or a time value. If passed a frequency (using either the ω or frequency keyword), the input to fn will be a phase value in radians, ranging from 0 to 2π. If no frequency is specified the value passed to f is the total time in seconds. Specifying phase (by the ϕ or phase keyword) will first add that value to the input. The phase is assumed to be in units of radians (but you can also pass degrees by using °).

source

If fn == randn no frequency or phase can be specified. Instead there is a signle keywored argument rng which allows you to specify the random number generator, which defaults to Random.GLOBAL_RNG.

source

SignalOperators.sink — Function.

sink([signal],[to=AxisArray];length,samplerate)

Creates a given type of object (to) from a signal. By default it is an AxisArray with time as the rows and channels as the columns. If a filename is specified for to, the signal is written to the given file. If given a type (e.g. Array) the signal is written to that type. The sample rate does not need to be specified, it will use either the sample rate of signal or a default sample rate (which raises a warning).

You can specify a length or samplerate for the signal when calling sink if it has yet to be defined.

source

SignalOperators.sink! — Function.

sink!(array,x;[samplerate],[offset])

Write samples of the signal x to array, starting from the sample after offset. If no sample rate has been specified for x you can specify it now, using samplerate (it will default to 44.1kHz).

source

Signal Inspection

SignalOperators.inflen — Constant.

inflen

Represents an infinite length. Proper overloads are defined to handle arithematic and ordering for the infinite value.

source

SignalOperators.duration — Function.

duration(x)

Return the duration of the signal in seconds, if known. May return missing or inflen. The value missing always denotes a finite, but unknown length.

source

SignalOperators.nsamples — Function.

nsamples(x)

Returns the number of samples in the signal, if known. May return missing or inflen. The value missing always denotes a finite, but unknown length.

source

SignalOperators.nchannels — Function.

nchannels(x)

Returns the number of channels in the signal.

source

SignalOperators.samplerate — Function.

samplerate(x)

Returns the sample rate of the signal (in Hz). May return missing if the sample rate is unknown.

source

Reformatting

SignalOperators.tosamplerate — Function.

tosamplerate(x,fs;blocksize)

Change the sample rate of x to the given sample rate fs. Functionally defined signals (e.g. signal(sin)) are resampled exactly: the function is simply called more times to generate more samples. Data-based signals (signal(rand(50,2))) are resampled using filtering. In this case you can use the keyword arugment blocksize to change the analysis window used. See filtersignal for more details.

source

SignalOperators.tochannels — Function.

tochannels(x,ch)

Force a signal to have ch number of channels, by mixing channels together or broadcasting a single channel over multiple channels.

source

SignalOperators.format — Function.

format(x,fs,ch)

Efficiently convert both the samplerate (fs) and channels ch of signal x. This selects an optimal ordering for tosamplerate and tochannels to avoid redundant computations.

source

SignalOperators.uniform — Function.

uniform(xs;channels=false)

Promote the sample rate (and optionally the number of channels) to be the sample highest sample rate (and optionally channel count) of the passed value xs, and iterable of signals.

Note

uniform rarely needs to be called directly. It is called implicitly, within the body of mapsignal for example.

source

Signal Manipulation

SignalOperators.until — Function.

until(x,time)

Create a signal of all samples of x up until time.

source

SignalOperators.after — Function.

after(x,time)

Create a signal of all samples of x after time.

source

SignalOperators.append — Function.

append(x,y,rest...)

Append a series of signals, one after the other.

source

SignalOperators.prepend — Function.

prepend(xs...)

Prepend the series of signals: prepend(xs...) = append(reverse(xs)...).

source

SignalOperators.pad — Function.

pad(x,padding)

Create a signal that appends an infinite number of values, padding, to x. The value padding can be a number or a function of a type (e.g. zero).

source

SignalOperators.filtersignal — Function.

filtersignal(x,h;[blocksize])

Apply the given filter h (from DSP) to signal x.

Blocksize

Blocksize determines the size of the buffer used when computing intermediate values of the filter. It defaults to 4096. It need not normally be adjusted.

source

SignalOperators.lowpass — Function.

lowpass(x,low;[order=5],[method=Butterworth(order)],[blocksize])

Apply a lowpass filter to x at the given cutoff frequency (low). See filtersignal for details on blocksize.

source

SignalOperators.highpass — Function.

highpass(x,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a highpass filter to x at the given cutoff frequency (low). See filtersignal for details on blocksize.

source

SignalOperators.bandpass — Function.

bandpass(x,low,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a bandpass filter to x at the given cutoff frequencies (low and high). See filtersignal for details on blocksize.

source

SignalOperators.bandstop — Function.

bandstop(x,low,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a bandstop filter to x at the given cutoff frequencies (low and high). See filtersignal for details on blocksize.

source

SignalOperators.normpower — Function.

normpower(x)

Divide all samples by the root-mean-squared value of the signal.

source

SignalOperators.mapsignal — Function.

mapsignal(fn,arguments...;padding,across_channels)

Apply fn across the samples of arguments, producing a signal of the output of fn. Shorter signals are padded to accomodate the longest finite-length signal. The function fn can return a single number or a tuple of numbers. In either case it is expected to be a type stable function.

Cross-channel functions

The function is normally broadcast across channels, but if you wish to treat each channel seperately you can set across_channels=true. In this case the inputs to fn will be tuples of all channel values for a given sample, and fn should return a type-stable tuple value. For exmpale, the following would swap the left and right channels.

x = rand(10,2)
swapped = mapsignal(x,across_channels=true) do (left,right)
    right,left
end

Padding

Padding determines how samples past the end of shorter signals are reported, it is set to a default specific fn using default_pad(fn). There is a fallback implementation which returns zero. You can pass a number or a function of a type (e.g. zero) to padding. The default for the four basic arithematic operators is their identity (one for * and zero for +).

To define a new default for a specific function, just create a new method of default_pad(fn)

myfun(x) = 2x + 3
SignalOperators.default_pad(myfun) = zero

source

SignalOperators.mix — Function.

mix(xs...)

Sum all signals together, using mapsignal

source

SignalOperators.amplify — Function.

amplify(xs...)

Multipy all signals by one another, using mapsignal

source

SignalOperators.addchannel — Function.

addchannel(xs...)

Concatenate the channels of all signals into one signal with sum(nchannels,xs) channels.

source

SignalOperators.channel — Function.

channel(x,n)

Select channel n of signal x, as a single channel signal.

source

SignalOperators.rampon — Function.

rampon(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds.

The function should be non-decreasing and should have a domain and range of [0,1]

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.rampoff — Function.

rampoff(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the offset of a signal, smoothly transitioning from frull amplitude to 0 amplitude over the course of len seconds.

The function should be non-decreasing and should have a domain and range of [0,1]

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.ramp — Function.

ramp(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset and offset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds at the start and from full to 0 amplitude over the course of len seconds.

The function should be non-decreasing and should have a domain and range of [0,1]

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.fadeto — Function.

fadeto(x,y,[len=10ms],[fn=x->sinpi(0.5x)])

Append x to y, with a smooth transition lasting len seconds fading from x to y (so the total length is duration(x) + duration(y) - len).

The function should be non-decreasing and should have a domain and range of [0,1]

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source