Reference

Reference

Signal Generation

SignalOperators.signal — Function.

signal(x,[samplerate])

Coerce x to be a signal, optionally specifying its sample rate (usually in Hz). All signal operators first call signal(x) for each argument. This means you only need to call signal when you want to pass additional arguments to it.

Note

If you pipe signal and pass a sample rate, you must specify the units of the sample rate (e.g. x |> signal(20Hz)). A unitless number is always interpreted as a constant, infinite-length signal (see below).

Note

If you are implementing signal for a custom signal, you will need to support the second argument of signal by specifying fs::Union{Number,Missing}=missing, or equivalent, as your second argument.

The type of objects that can be coerced to signals are as follows.

source

Existing signals

Any existing signal just returns itself from signal. If a sample rate is specified it will be set if x has an unknown sample rate. If it has a known sample rate and doesn't match samplerate(x) an error will be thrown. If you want to change the sample rate of a signal use tosamplerate.

source

Numbers

Numbers can be treated as infinite length, constant signals of unknown sample rate.

source

Arrays

Arrays can be treated as signals. The first dimension is time, the second channels.

AxisArrays, if they have an axis labeled time and one or zero additional axes, can be treated as a signal. The time dimension must be represented using on object with the step function defined (e.g. any AbstractRange).

source

Functions

signal(fn,[samplerate];[ω/frequency],[ϕ/phase])

Functions can define infinite length signals of known or unknown sample rate. The function fn can either return a number or, for multi-channel signals, a tuple of values.

The input to fn is either a phase value or a time value. If a frequency is specified (using either the ω or frequency keyword), the input to fn will be a phase value in radians, ranging from 0 to 2π. If no frequency is specified the value passed to fn is the time in seconds. Specifying phase (by the ϕ or phase keyword) will first add that value to the input before passing it to fn. When frequency is specified, the phase is assumed to be in units of radians (but you can also pass degrees by using ° or a unit of time (e.g. s for seconds)). When frequency is not specified the phase is assumed to be in units of seconds.

source

If fn == randn no frequency or phase can be specified. Instead there is a single keyword argument, rng, which allows you to specify the random number generator; rng defaults to Random.GLOBAL_RNG.

source

SignalOperators.sink — Function.

sink([signal],[to=AxisArray];duration,samplerate)

Creates a given type of object (to) from a signal. By default it is an AxisArray with time as the rows and channels as the columns. If a filename is specified for to, the signal is written to the given file. If given a type (e.g. Array) the signal is written to a value of that type.

Sample Rate

The sample rate does not need to be specified, it will use either the sample rate of signal or a default sample rate (which raises a warning). If specified, the given sample rate is passed to signal when coercing the input to a signal.

Duration

You can limit the output of the given signal to the specified duration. If this duration exceedes the duration of the passed signal an error will be thrown.

source

SignalOperators.sink! — Function.

sink!(array,x;[samplerate])

Write size(array,1) samples of signal x to array. If no sample rate has been specified for x you can specify it now, using samplerate (it will default to 44.1kHz).

source

Signal Inspection

SignalOperators.inflen — Constant.

inflen

Represents an infinite length. Proper overloads are defined to handle arithmetic and ordering for the infinite value.

Missing values

For the purposes of SignalOperators missing values are considered to be unknown, but of finite length. For example: inflen * missing == inflen.

source

SignalOperators.duration — Function.

duration(x)

Return the duration of the signal in seconds, if known. May return missing or inflen. The value missing always denotes a finite but unknown length.

Note

If your are implementing a custom signal, you need not normally define duration as it will be computed from nsamples and samplerate. However, if one or both of these is missing and you want duartion to return a non-missing value, you can define custom method of duration.

source

SignalOperators.nsamples — Function.

nsamples(x)

Returns the number of samples in the signal, if known. May return missing or inflen. The value missing always denotes a finite but unknown length.

Note

The return value of nsamples for a block (see custom signals must be a non-missing, finite value.

source

SignalOperators.nchannels — Function.

nchannels(x)

Returns the number of channels in the signal.

source

SignalOperators.samplerate — Function.

samplerate(x)

Returns the sample rate of the signal (in Hertz). May return missing if the sample rate is unknown.

source

SignalOperators.channel_eltype — Function.

channel_eltype(x)

Returns the element type of an individual channel of a signal (e.g. Float64).

Note

channel_eltype and eltype are, in most cases, the same, but not necesarilly so.

source

Signal Reformatting

SignalOperators.tosamplerate — Function.

tosamplerate(x,fs;blocksize)

Change the sample rate of x to the given sample rate fs. The underlying implementation depends on whether the input is a computed or data signal, as determined by EvalTrait.

Computed signals (e.g. signal(sin)) are resampled exactly: the result is simply computed for more time points or fewer time points, so as to generate the appropriate number of samples.

Data-based signals (signal(rand(50,2))) are resampled using filtering (akin to DSP.resample). In this case you can use the keyword arugment blocksize to change the analysis window used. See filtersignal for more details. Setting blocksize for a computed signal will succeed, but different blocksize values have no effect on the underlying implementation.

Implementation

You need only implement this function for custom signals for particular scenarios, described below.

Custom Computed Signals

If you implement a new sigal type that is a computed signal, you must implement tosamplerate with the following type signature.

function tosamplerate(x::MyCustomSignal,s::IsSignal{<:Any,<:Number},
    c::ComputedSignal,samplerate;blocksize)

    ## ...
end

The result should be a new version of the computed signal with the given sample rate.

Handling missing sample rates

If you implement a new signal type that can handle missing sample rate values, you will need to implement the following version of tosamplerate so that a known sample rate can be applied to a signal with a missing sample rate.

function tosamplerate(x::MyCustomSignal,s::IsSignal{<:Any,Missing},
    evaltrait,samplerate;blocksize)

    ## ...
end

The result should be a new version of the signal with the specified sample rate.

source

SignalOperators.tochannels — Function.

tochannels(x,ch)

Force a signal to have ch number of channels, by mixing channels together or broadcasting a single channel over multiple channels.

source

SignalOperators.toeltype — Function.

toeltype(x,T)

Converts individual samples in signal x to type T.

source

SignalOperators.format — Function.

format(x,fs,ch)

Efficiently convert both the samplerate (fs) and channels ch of signal x. This selects an optimal ordering for tosamplerate and tochannels to avoid redundant computations.

source

SignalOperators.uniform — Function.

uniform(xs;channels=false)

Promote the sample rate (and optionally the number of channels) to be the highest sample rate (and optionally highest channel count) of the iterable of signals xs.

Note

uniform rarely needs to be called directly. It is called implicitly on all passed signals, within the body of operators such as mapsignal.

source

Signal Operators

SignalOperators.until — Function.

until(x,time)

Create a signal of all samples of x up until and including time.

source

SignalOperators.after — Function.

after(x,time)

Create a signal of all samples of x after time.

source

SignalOperators.append — Function.

append(x,y,...)

Append a series of signals, one after the other.

source

SignalOperators.prepend — Function.

prepend(x,y,...)

Prepend the series of signals: prepend(xs...) is equivalent to append(reverse(xs)...).

source

SignalOperators.pad — Function.

pad(x,padding)

Create a signal that appends an infinite number of values, padding, to x. The value padding can be:

a number

a tuple or vector

a type function: a one argument function of the channel_eltype of x

a value function: a one argument function of the signal x for which SignalOperators.valuefunction(padding) == true.

an indexing function: a three argument function following the same type signature as getindex for two dimensional arrays.

If the signal is already infinitely long (e.g. a previoulsy padded signal), pad has no effect.

If padding is a number it is used as the value for all samples and channels past the end of x.

If padding is a tuple or vector it is the value for all samples past the end of x.

If padding is a type function it is passed the channel_eltype of the signal and the resulting value is used as the value for all samples past the end of x. Examples include zero and one

If padding is a value function it is passed x just before padding during sink begins and it should return a tuple of channel_eltype(x) values. This value is repeated for the remaining samples. It is generally only useful when x is an AbstractArray.

If padding is an indexing function (it accepts 3 arguments) it will be used to retrieve samples from the signal x assuming it conforms to the AbstractArray interface, with the first index being samples and the second channels. If the sample index goes past the bounds of the array, it should be transformed to an index within the range of that array. Note that such padding functions only work on signals that are also AbstractArray objects. You can always generate an array from a given signal by first passing it through sink or sink!.

See also

cycle mirror lastsample valuefunction

source

SignalOperators.mirror — Function.

mirror(x,i,j)

An indexing function which mirrors the indices when i > size(x,1). This means that past the end of the signal x, the signal first repeats with samples in reverse order, then repeats in the original order, so on and so forth. It can be passed as the second argument to pad.

source

SignalOperators.cycle — Function.

cycle(x,i,j)

An indexing function which wraps index i using mod, thus repeating the signal when i > size(x,1). It can be passed as the second argument to pad.

source

SignalOperators.lastsample — Function.

lastsample

When passed as an argument to pad, allows padding using the last sample of a signal. You cannot use this function in other contexts, and it will normally throw an error. See pad.

source

SignalOperators.valuefunction — Function.

SignalOperators.valuefunction(fn)

Returns true if fn should be treated as a value function. See pad. If you wish your own function to be a value function, you can do this as follows.

SignalOperators.valuefunction(::typeof(myfun)) = true

source

SignalOperators.filtersignal — Function.

filtersignal(x,h;[blocksize])

Apply the given filter h (from DSP) to signal x.

Blocksize

Blocksize determines the size of the buffer used when computing intermediate values of the filter. It defaults to 4096. It need not normally be adjusted.

source

SignalOperators.lowpass — Function.

lowpass(x,low;[order=5],[method=Butterworth(order)],[blocksize])

Apply a lowpass filter to x at the given cutoff frequency (low). See filtersignal for details on blocksize.

source

SignalOperators.highpass — Function.

highpass(x,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a highpass filter to x at the given cutoff frequency (low). See filtersignal for details on blocksize.

source

SignalOperators.bandpass — Function.

bandpass(x,low,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a bandpass filter to x at the given cutoff frequencies (low and high). See filtersignal for details on blocksize.

source

SignalOperators.bandstop — Function.

bandstop(x,low,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a bandstop filter to x at the given cutoff frequencies (low and high). See filtersignal for details on blocksize.

source

SignalOperators.normpower — Function.

normpower(x)

Return a signal with normalized power. That is, divide all samples by the root-mean-squared value of the entire signal.

source

SignalOperators.mapsignal — Function.

mapsignal(fn,arguments...;padding,bychannel)

Apply fn across the samples of arguments, producing a signal of the output of fn. Shorter signals are padded to accommodate the longest finite-length signal. The function fn should treat each argument as a single number and return a single number. This operation is broadcast across all channels of the input. It is expected to be a type stable function.

Normally the signals are first promoted to have the same samle rate and the same number of channels using uniform (with channels=true).

Cross-channel functions

The function fn is normally broadcast across channels, but if you wish to treat each channel separately you can set bychannel=false. In this case the inputs to fn will be indexable objects (tuples or arrays) of all channel values for a given sample, and fn should return a type-stable tuple value (for a multi-channel or single-channel result) or a number (for a single-channel result only). For example, the following would swap the left and right channels.

x = rand(10,2)
swapped = mapsignal(x,bychannel=false) do val
    val[2],val[1]
end

When bychannel=false the channels of each signal are not promoted:

Padding

Padding determines how samples past the end of shorter signals are reported. The value of padding is passd to pad. Its default value is determined by the value of fn. The default value for the four basic arithmetic operators is their identity (one for * and zero for +). These defaults are set on the basis of fn using default_pad(fn). A fallback implementation of default_pad returns zero.

To define a new default for a specific function, just create a new method of default_pad(fn)

myfun(x) = 2x + 3
SignalOperators.default_pad(::typeof(myfun)) = one

source

SignalOperators.mix — Function.

mix(xs...)

Sum all signals together, using mapsignal

source

SignalOperators.amplify — Function.

amplify(xs...)

Find the product, on a per-sample basis, for all signals xs using mapsignal.

source

SignalOperators.addchannel — Function.

addchannel(xs...)

Concatenate the channels of all signals into one signal, using mapsignal. This will result in a signal with sum(nchannels,xs) channels.

source

SignalOperators.channel — Function.

channel(x,n)

Select channel n of signal x, as a single-channel signal, using mapsignal.

source

SignalOperators.rampon — Function.

rampon(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds.

The function determines the shape of the ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.rampoff — Function.

rampoff(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the offset of a signal, smoothly transitioning from full amplitude to 0 amplitude over the course of len seconds.

The function determines the shape of the ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.ramp — Function.

ramp(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset and offset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds at the start and from full to 0 amplitude over the course of len seconds.

The function determines the shape of the ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.fadeto — Function.

fadeto(x,y,[len=10ms],[fn=x->sinpi(0.5x)])

Append x to y, with a smooth transition lasting len seconds fading from x to y (so the total length is duration(x) + duration(y) - len).

This fade is accomplished with a rampoff of x and a rampon for y. fn should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

Custom Signals

SignalOperators.SignalTrait — Function.

SiganlOperators.SignalTrait(::Type{T}) where T

Returns either nothing if the type T should not be considered a signal (the default) or IsSignal to indicate the signal format for this signal.

source

SignalOperators.IsSignal — Type.

SignalOperators.IsSignal{T,Fs,L}

Represents the format of a signal type with three type parameters:

T - The channel_eltype of the signal.

Fs - The type of the samplerate. It should be either Float64 or Missing.

L - The type of the length of the signal. It should be either

InfiniteLength, Missing or Int.

source

SignalOperators.EvalTrait — Function.

SiganlOperators.EvalTrait(x)

Indicates whether the signal is a DataSignal or ComputedSignal. Data signals represent samples concretely as a set of samples. Examples include arrays and numbers. Data signals generally return themselves, or some wrapper type when sink is called on them. Computed signals are any signal that invovles some intermediate computation, in which samples must be computued on the fly. Calls to sink on a computed signal results in some new, data signal. Most signals returned by a signal operator are computed signals.

Computed signals have the extra responsibility of implementing tosamplerate

source

SignalOperators.nextblock — Function.

SignalOperators.nextblock(x,maxlength,skip,[block])

Retrieve the next block of samples for signal x. The final, fourth argument is optional. If it is left out, nextblock returns the first block of the signal. The resulting block must has no more than maxlength samples, but may have fewer samples than that; it should not have zero samples unless maxlength == 0. If skip == true, it is guaranted that sample will never be called on the returned block. The value of skip is true, for example, when skipping blocks during a call to after).

source

SignalOperators.sample — Function.

SignalOperators.sample(x,block,i)

Retrieves the sample at index i of the given block of signal x. A sample is one or more channels of chnanle_eltypep(x) values. The return value should be an indexable object (e.g. a number, tuple or array) of these channel values. This method should be implemented by blocks of custom signals.

source