Reference

Signal Generation

Missing docstring.

Missing docstring for signal. Check Documenter's build log for details.

SignalOperators.sink — Function

sink([signal],[to];duration,framerate)

Creates a given type of object (to) from a signal. By default the type of the resulting sink is determined by the type of the underlying data of the signal: e.g. if x is a SampleBuf object then sink(Mix(x,2)) is also a SampleBuf object. If there is no underlying data (Signal(sin) |> sink) then a Tuple of an array and the framerate is returned.

Keyword arguments

Frame Rate

The frame rate does not need to be specified, it will use either the frame rate of Signal or a default frame rate (which raises a warning). If specified, the given frame rate is passed to Signal when coercing the input to a signal.

Duration

You can limit the output of the given signal to the specified duration. If this duration exceedes the duration of the passed signal an error will be thrown.

Values for to

Type

If to is a type (e.g. Array) the signal is written to a value of that type.

Filename

If to is a string, it is assumed to describe the name of a file to which the signal will be written. You will need to call import or using on an appropriate backend for writing to the given file type.

Available backends include the following pacakges

WAV

LibSndFile

SignalOperators.sink! — Function

sink!(array,x;[framerate])

Write size(array,1) frames of signal x to array. If no frame rate has been specified for x you can specify it now, using framerate (it will default to 44.1kHz).

Signal Inspection

SignalOperators.inflen — Constant

inflen

Represents an infinite length. Proper overloads are defined to handle arithmetic and ordering for the infinite value.

Missing values

For the purposes of SignalOperators missing values are considered to be unknown, but of finite length. For example: inflen * missing == inflen.

SignalOperators.duration — Function

duration(x)

Return the duration of the signal in seconds, if known. May return missing or inflen. The value missing always denotes a finite but unknown length.

Note

If your are implementing a custom signal, you need not normally define duration as it will be computed from nframes and framerate. However, if one or both of these is missing and you want duartion to return a non-missing value, you can define custom method of duration.

SignalOperators.nframes — Function

nframes(x)

Returns the number of frames in the signal, if known. May return missing or inflen. The value missing always denotes a finite but unknown length.

Note

The return value of nframes for a block (see custom signals must be a non-missing, finite value.

SignalOperators.nchannels — Function

nchannels(x)

Returns the number of channels in the signal.

SignalOperators.framerate — Function

framerate(x)

Returns the frame rate of the signal (in Hertz). May return missing if the frame rate is unknown.

SignalOperators.channel_eltype — Function

channel_eltype(x)

Returns the element type of an individual channel of a signal (e.g. Float64).

Note

channel_eltype and eltype are, in most cases, the same, but not necesarilly so.

Signal Operators

Basic Operators

SignalOperators.Until — Function

Until(x,time)

Create a signal of all frames of x up Until and including time.

SignalOperators.until — Function

until(x,time)

Equivalent to sink(Until(x,time))

See also

Until

SignalOperators.After — Function

After(x,time)

Create a signal of all frames of x After time.

SignalOperators.after — Function

after(x,time)

Equivalent to sink(After(x,time))

See also

After

SignalOperators.Append — Function

Append(x,y,...)

Append a series of signals, one after the other.

SignalOperators.append — Function

append(x,y,...)

Equivalent to sink(Append(x,y,...))

See also

Append

SignalOperators.Prepend — Function

Prepend(x,y,...)

Prepend the series of signals: Prepend(xs...) is equivalent to Append(reverse(xs)...).

Missing docstring.

Missing docstring for preprend. Check Documenter's build log for details.

SignalOperators.Pad — Function

Pad(x,padding)

Create a signal that appends an infinite number of values, padding, to x. The value padding can be:

a number

a tuple or vector

a type function: a one argument function of the channel_eltype of x

a value function: a one argument function of the signal x for which SignalOperators.valuefunction(padding) == true.

an indexing function: a three argument function following the same type signature as getindex for two dimensional arrays.

If the signal is already infinitely long (e.g. a previoulsy padded signal), Pad has no effect.

If padding is a number it is used as the value for all frames and channels past the end of x.

If padding is a tuple or vector it is the value for all frames past the end of x.

If padding is a type function it is passed the channel_eltype of the signal and the resulting value is used as the value for all frames past the end of x. Examples include zero and one

If padding is a value function it is passed x just before padding during sink begins and it should return a tuple of channel_eltype(x) values. This value is repeated for the remaining frames. It is generally only useful when x is an AbstractArray.

If padding is an indexing function (it accepts 3 arguments) it will be used to retrieve frames from the signal x assuming it conforms to the AbstractArray interface, with the first index being frames and the second channels. If the frame index goes past the bounds of the array, it should be transformed to an index within the range of that array. Note that such padding functions only work on signals that are also AbstractArray objects. You can always generate an array from a given signal by first passing it through sink or sink!.

See also

cycle mirror lastframe valuefunction

SignalOperators.mirror — Function

mirror(x,i,j)

An indexing function which mirrors the indices when i > size(x,1). This means that past the end of the signal x, the signal first repeats with frames in reverse order, then repeats in the original order, so on and so forth. It can be passed as the second argument to Pad.

SignalOperators.cycle — Function

cycle(x,i,j)

An indexing function which wraps index i using mod, thus repeating the signal when i > size(x,1). It can be passed as the second argument to Pad.

SignalOperators.lastframe — Function

lastframe

When passed as an argument to Pad, allows padding using the last frame of a signal. You cannot use this function in other contexts, and it will normally throw an error. See Pad.

SignalOperators.valuefunction — Function

SignalOperators.valuefunction(fn)

Returns true if fn should be treated as a value function. See Pad. If you wish your own function to be a value function, you can do this as follows.

SignalOperators.valuefunction(::typeof(myfun)) = true

Mapping Operators

SignalOperators.Filt — Function

Filt(x,::Type{<:FilterType},bounds...;method=Butterworth(order),order=5,
     blocksize=4096)

Apply the given filter type (e.g. Lowpass) using the given method to design the filter coefficients. The type is specified as per the types from DSP

Filt(x,h;[blocksize=4096])

Apply the given digital filter h (from DSP) to signal x.

Blocksize

Blocksize determines the size of the buffer used when computing intermediate values of the filter. It need not normally be adjusted, though changing it can alter how efficient filter application is.

Note

The non-lazy version of Filt is filt from the DSP package. Proper methods have been defined such that it should be possible to call filt on a signal and get a signal back.

The argument order for filt follows a different convention, with x coming after the filter specification. In contrast, Filt uses the convention of keeping x as the first argument to make piping possible.

SignalOperators.Normpower — Function

Normpower(x)

Return a signal with normalized power. That is, divide all frames by the root-mean-squared value of the entire signal.

SignalOperators.MapSignal — Type

MapSignal(fn,arguments...;padding,bychannel)

Apply fn across the frames of arguments, producing a signal of the output of fn. Shorter signals are padded to accommodate the longest finite-length signal. The function fn should treat each argument as a single number and return a single number. This operation is broadcast across all channels of the input. It is expected to be a type stable function.

Normally the signals are first promoted to have the same samle rate and the same number of channels using Uniform (with channels=true).

Cross-channel functions

The function fn is normally broadcast across channels, but if you wish to treat each channel separately you can set bychannel=false. In this case the inputs to fn will be indexable objects (tuples or arrays) of all channel values for a given frame, and fn should return a type-stable tuple value (for a multi-channel or single-channel result) or a number (for a single-channel result only). For example, the following would swap the left and right channels.

x = rand(10,2)
swapped = MapSignal(x,bychannel=false) do val
    val[2],val[1]
end

When bychannel=false the channels of each signal are not promoted:

Padding

Padding determines how frames past the end of shorter signals are reported. The value of padding is passd to Pad. Its default value is determined by the value of fn. The default value for the four basic arithmetic operators is their identity (one for * and zero for +). These defaults are set on the basis of fn using default_pad(fn). A fallback implementation of default_pad returns zero.

To define a new default for a specific function, just create a new method of default_pad(fn)

myfun(x) = 2x + 3
SignalOperators.default_pad(::typeof(myfun)) = one

SignalOperators.mapsignal — Function

mapsignal(fn,args...;padding,bychannel)

Equivalent to sink(MapSignal(fn,args...;padding,bychannel))

See also

MapSignal

SignalOperators.Mix — Function

Mix(xs...)

Sum all signals together, using MapSignal

SignalOperators.mix — Function

mix(xs...)

Equivalent to sink(Mix(xs...))

See also

Mix

SignalOperators.Amplify — Function

Amplify(xs...)

Find the product, on a per-frame basis, for all signals xs using MapSignal.

SignalOperators.amplify — Function

amplify(xs...)

Equivalent to sink(Amplify(xs...))

See also

Amplify

SignalOperators.AddChannel — Function

AddChannel(xs...)

Concatenate the channels of all signals into one signal, using MapSignal. This will result in a signal with sum(nchannels,xs) channels.

SignalOperators.addchannel — Function

addchannel(xs...)

Equivalent to sink(AddChannel(xs...))

See also

AddChannel

SignalOperators.SelectChannel — Function

SelectChannel(x,n)

Select channel n of signal x, as a single-channel signal, using MapSignal.

SignalOperators.selectchannel — Function

selectchannel(xs...)

Equivalent to sink(SelectChannel(xs...))

See also

SelectChannel

Ramping Operators

SignalOperators.RampOn — Function

RampOn(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds.

The function determines the shape of the Ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

SignalOperators.rampon — Function

rampon(x,[len],[fn])

Equivalent to sink(RampOn(x,[len],[fn]))

See also

RampOn

SignalOperators.RampOff — Function

RampOff(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the offset of a signal, smoothly transitioning from full amplitude to 0 amplitude over the course of len seconds.

The function determines the shape of the Ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

SignalOperators.rampoff — Function

rampoff(x,[len],[fn])

Equivalent to sink(RampOff(x,[len],[fn]))

See also

RampOff

SignalOperators.Ramp — Function

Ramp(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset and offset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds at the start and from full to 0 amplitude over the course of len seconds.

The function determines the shape of the Ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

SignalOperators.ramp — Function

ramp(x,[len],[fn])

Equivalent to sink(Ramp(x,[len],[fn]))

See also

Ramp

SignalOperators.FadeTo — Function

FadeTo(x,y,[len=10ms],[fn=x->sinpi(0.5x)])

Append x to y, with a smooth transition lasting len seconds fading from x to y (so the total length is duration(x) + duration(y) - len).

This fade is accomplished with a RampOff of x and a RampOn for y. fn should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

SignalOperators.fadeto — Function

fadeto(x,y,[len],[fn])

Equivalent to sink(FadeTo(x,[len],[fn]))

See also

FadeTo

Reformatting Operators

SignalOperators.ToFramerate — Function

ToFramerate(x,fs;blocksize)

Change the frame rate of x to the given frame rate fs. The underlying implementation depends on whether the input is a computed or data signal, as determined by EvalTrait.

Computed signals (e.g. Signal(sin)) are resampled exactly: the result is simply computed for more time points or fewer time points, so as to generate the appropriate number of frames.

Data-based signals (Signal(rand(50,2))) are resampled using filtering (akin to DSP.resample). In this case you can use the keyword arugment blocksize to change the analysis window used. See Filt for more details. Setting blocksize for a computed signal will succeed, but different blocksize values have no effect on the underlying implementation.

Implementation

You need only implement this function for custom signals for particular scenarios, described below.

Custom Computed Signals

If you implement a new sigal type that is a computed signal, you must implement ToFramerate with the following type signature.

function ToFramerate(x::MyCustomSignal,s::IsSignal{<:Any,<:Number},
    c::ComputedSignal,framerate;blocksize)

    ## ...
end

The result should be a new version of the computed signal with the given frame rate.

Handling missing frame rates

If you implement a new signal type that can handle missing frame rate values, you will need to implement the following version of ToFramerate so that a known frame rate can be applied to a signal with a missing frame rate.

function ToFramerate(x::MyCustomSignal,s::IsSignal{<:Any,Missing},
    evaltrait,framerate;blocksize)

    ## ...
end

The result should be a new version of the signal with the specified frame rate.

SignalOperators.toframerate — Function

toframerate(x,fs;blocksize)

Equivalent to sink(ToFramerate(x,fs;blocksize=blocksize))

See also

ToFramerate

SignalOperators.ToChannels — Function

ToChannels(x,ch)

Force a signal to have ch number of channels, by Mixing channels together or broadcasting a single channel over multiple channels.

SignalOperators.tochannels — Function

tochannels(x,ch)

Equivalent to sink(ToChannels(x,ch))

See also

ToFramerate

SignalOperators.ToEltype — Function

ToEltype(x,T)

Converts individual frames in signal x to type T.

SignalOperators.toeltype — Function

toeltype(x,T)

Equivalent to sink(ToEltype(x,T))

See also

ToEltype

SignalOperators.Format — Function

Format(x,fs,ch)

Efficiently convert both the framerate (fs) and channels ch of signal x. This selects an optimal ordering for ToFramerate and ToChannels to avoid redundant computations.

SignalOperators.format — Function

format(x,fs,ch)

Equivalent to sink(Format(x,fs,ch))

See also

Format

SignalOperators.Uniform — Function

Uniform(xs;channels=false)

Promote the frame rate (and optionally the number of channels) to be the highest frame rate (and optionally highest channel count) of the iterable of signals xs.

Note

Uniform rarely needs to be called directly. It is called implicitly on all passed signals, within the body of operators such as MapSignal.

Custom Signals

SignalOperators.SignalTrait — Function

SiganlOperators.SignalTrait(::Type{T}) where T

Returns either nothing if the type T should not be considered a signal (the default) or IsSignal to indicate the signal Format for this signal.

SignalOperators.IsSignal — Type

SignalOperators.IsSignal{T,Fs,L}

Represents the Format of a signal type with three type parameters:

T - The channel_eltype of the signal.

Fs - The type of the framerate. It should be either Float64 or Missing.

L - The type of the length of the signal. It should be either

InfiniteLength, Missing or Int.

SignalOperators.EvalTrait — Function

SiganlOperators.EvalTrait(x)

Indicates whether the signal is a DataSignal or ComputedSignal. Data signals represent frames concretely as a set of frames. Examples include arrays and numbers. Data signals generally return themselves, or some wrapper type when sink is called on them. Computed signals are any signal that invovles some intermediate computation, in which frames must be computued on the fly. Calls to sink on a computed signal results in some new, data signal. Most signals returned by a signal operator are computed signals.

Computed signals have the extra responsibility of implementing ToFramerate

SignalOperators.nextblock — Function

SignalOperators.nextblock(x,maxlength,skip,[block])

Retrieve the next block of frames for signal x. The final, fourth argument is optional. If it is left out, nextblock returns the first block of the signal. The resulting block must has no more than maxlength frames, but may have fewer frames than that; it should not have zero frames unless maxlength == 0. If skip == true, it is guaranted that frame will never be called on the returned block. The value of skip is true, for example, when skipping blocks during a call to After).

SignalOperators.frame — Function

SignalOperators.frame(x,block,i)

Retrieves the frame at index i of the given block of signal x. A frame is one or more channels of channel_eltype(x) values. The return value should be an indexable object (e.g. a number, tuple or array) of these channel values. This method should be implemented by blocks of custom signals.

SignalOperators.timeslice — Function

SignalOperators.timeslice(x::AbstractArray,indices)

Extract the slice of x with the given time indices.

Custom signals can implement this method if the signal is an AbstractArray allowing the use of a fallback implementation of SignalOperators.nextblock.

SignalOperators.ArrayBlock — Type

ArrayBlock{A,S}(data::A,state::S)

A straightforward implementation of blocks as an array and a custom state. The array allows a generic implementation of nframes and SignalOperators.frame. The fields of this struct are data and state.

Custom signals can return an ArrayBlock from SignalOperators.nextblock to allow for fallback implementations of nframes and SignalOperators.frame.

Custom Sinks

SignalOperators.initsink — Function

SignalOperators.initsink(x,::Type{T},len)

Initialize an object of type T so that it can store the first len frames of signal x.

If you wish an object to serve as a custom sink you can implement this method. You should probably use nchannels and channel_eltype of x to determine how to initialize the object.

SignalOperators.sink_helper! — Function

SignalOperators.sink_helper!(result,written,x,block)

Write the given block of frames from signal x to result given that a total of written frames have already been written to the result.

This method should be fast: i.e. a for loop using @simd and @inbounds. It should call nframes and SignalOperators.frame on the block to write the frames. Do not call frame more than once for each index of the block.

```