Reference

Reference

Signal Generation

SignalOperators.signal — Function.

signal(x,[samplerate])

Coerce x to be a signal, optionally specifying its sample rate (usually in Hz). All signal operators first call signal(x) for each argument. This means you only need to call signal when you want to pass additional arguments to it.

Note

If you pipe signal and pass a sample rate, you must specify the units of the sample rate (e.g. x |> signal(20Hz)). A unitless number is always interpreted as a constant, infinite-length signal (see below).

The type of objects that can be coerced to signals are as follows.

source

Existing signals

Any existing signal just returns itself from signal. If a sample rate is specified it will be set if x has an unknown sample rate. If it has a known sample rate and doesn't match samplerate(x) an error will be thrown. If you want to change the sample rate of a signal use tosamplerate.

source

Numbers

Numbers can be treated as infinite length, constant signals of unknown sample rate.

source

Arrays

Arrays can be treated as signals. The first dimension is time, the second channels.

AxisArrays, if they have an axis labeled time and one or zero additional axes, can be treated as a signal. The time dimension must be represented using on object with the step function defined (e.g. any AbstractRange).

source

Functions

signal(fn,[samplerate];[ω/frequency],[ϕ/phase])

Functions can define infinite length signals of known or unknown sample rate. The function fn can either return a number or, for multi-channel signals, a tuple of values.

The input to fn is either a phase value or a time value. If a frequency is specified (using either the ω or frequency keyword), the input to fn will be a phase value in radians, ranging from 0 to 2π. If no frequency is specified the value passed to fn is the time in seconds. Specifying phase (by the ϕ or phase keyword) will first add that value to the input before passing it to fn. When frequency is specified, the phase is assumed to be in units of radians (but you can also pass degrees by using ° or a unit of time (e.g. s for seconds)). When frequency is not specified the phase is assumed to be in units of seconds.

source

If fn == randn no frequency or phase can be specified. Instead there is a single keyword argument, rng, which allows you to specify the random number generator; rng defaults to Random.GLOBAL_RNG.

source

SignalOperators.sink — Function.

sink([signal],[to=AxisArray];duration,samplerate)

Creates a given type of object (to) from a signal. By default it is an AxisArray with time as the rows and channels as the columns. If a filename is specified for to, the signal is written to the given file. If given a type (e.g. Array) the signal is written to a value of that type.

Sample Rate

The sample rate does not need to be specified, it will use either the sample rate of signal or a default sample rate (which raises a warning). If specified, the given sample rate is passed to signal when coercing the input to a signal.

Duration

You can limit the output of the given signal to the specified duration. If this duration exceedes the duration of the passed signal an error will be thrown.

source

SignalOperators.sink! — Function.

sink!(array,x;[samplerate],[offset])

Write size(array,1) samples of signal x to array, starting from the sample after offset. If no sample rate has been specified for x you can specify it now, using samplerate (it will default to 44.1kHz).

source

Signal Inspection

SignalOperators.inflen — Constant.

inflen

Represents an infinite length. Proper overloads are defined to handle arithmetic and ordering for the infinite value.

Missing values

For the purposes of SignalOperators missing values are considered to be unknown, but of finite length. For example: inflen * missing == inflen.

source

SignalOperators.duration — Function.

duration(x)

Return the duration of the signal in seconds, if known. May return missing or inflen. The value missing always denotes a finite but unknown length.

source

SignalOperators.nsamples — Function.

nsamples(x)

Returns the number of samples in the signal, if known. May return missing or inflen. The value missing always denotes a finite but unknown length.

source

SignalOperators.nchannels — Function.

nchannels(x)

Returns the number of channels in the signal.

source

SignalOperators.samplerate — Function.

samplerate(x)

Returns the sample rate of the signal (in Hertz). May return missing if the sample rate is unknown.

source

Signal Reformatting

SignalOperators.tosamplerate — Function.

tosamplerate(x,fs;blocksize)

Change the sample rate of x to the given sample rate fs. The underlying implementation depends on whether the signal is defined by a function or by actual data (samples).

Functionally defined signals (e.g. signal(sin)) are resampled exactly: the function is simply called more times or fewer times, so as to generate the appropriate number of samples.

Data-based signals (signal(rand(50,2))) are resampled using filtering (akin to DSP.resample). In this case you can use the keyword arugment blocksize to change the analysis window used. See filtersignal for more details. Setting blocksize for a functionally defined signal will succeed, but different blocksize values have no effect on the underlying implementation.

source

SignalOperators.tochannels — Function.

tochannels(x,ch)

Force a signal to have ch number of channels, by mixing channels together or broadcasting a single channel over multiple channels.

source

SignalOperators.toeltype — Function.

toeltype(x,T)

Converts individual samples in signal x to type T.

source

SignalOperators.format — Function.

format(x,fs,ch)

Efficiently convert both the samplerate (fs) and channels ch of signal x. This selects an optimal ordering for tosamplerate and tochannels to avoid redundant computations.

source

SignalOperators.uniform — Function.

uniform(xs;channels=false)

Promote the sample rate (and optionally the number of channels) to be the highest sample rate (and optionally highest channel count) of the iterable of signals xs.

Note

uniform rarely needs to be called directly. It is called implicitly on all passed signals, within the body of operators such as mapsignal.

source

Signal Operators

SignalOperators.until — Function.

until(x,time)

Create a signal of all samples of x up until and including time.

source

SignalOperators.after — Function.

after(x,time)

Create a signal of all samples of x after time.

source

SignalOperators.append — Function.

append(x,y,...)

Append a series of signals, one after the other.

source

SignalOperators.prepend — Function.

prepend(x,y,...)

Prepend the series of signals: prepend(xs...) is equivalent to append(reverse(xs)...).

source

SignalOperators.pad — Function.

pad(x,padding)

Create a signal that appends an infinite number of values, padding, to x. The value padding can be a number or it can be a function of a type (e.g. zero).

source

SignalOperators.filtersignal — Function.

filtersignal(x,h;[blocksize])

Apply the given filter h (from DSP) to signal x.

Blocksize

Blocksize determines the size of the buffer used when computing intermediate values of the filter. It defaults to 4096. It need not normally be adjusted.

source

SignalOperators.lowpass — Function.

lowpass(x,low;[order=5],[method=Butterworth(order)],[blocksize])

Apply a lowpass filter to x at the given cutoff frequency (low). See filtersignal for details on blocksize.

source

SignalOperators.highpass — Function.

highpass(x,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a highpass filter to x at the given cutoff frequency (low). See filtersignal for details on blocksize.

source

SignalOperators.bandpass — Function.

bandpass(x,low,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a bandpass filter to x at the given cutoff frequencies (low and high). See filtersignal for details on blocksize.

source

SignalOperators.bandstop — Function.

bandstop(x,low,high;[order=5],[method=Butterworth(order)],[blocksize])

Apply a bandstop filter to x at the given cutoff frequencies (low and high). See filtersignal for details on blocksize.

source

SignalOperators.normpower — Function.

normpower(x)

Return a signal with normalized power. That is, divide all samples by the root-mean-squared value of the entire signal.

source

SignalOperators.mapsignal — Function.

mapsignal(fn,arguments...;padding,bychannel)

Apply fn across the samples of arguments, producing a signal of the output of fn. Shorter signals are padded to accommodate the longest finite-length signal. The function fn should treat each argument as a single number and return a single number. This operation is broadcast across all channels of the input. It is expected to be a type stable function.

Normally the signals are first promoted to have the same samle rate and the same number of channels using unify (with channels=true).

Cross-channel functions

The function fn is normally broadcast across channels, but if you wish to treat each channel separately you can set bychannel=false. In this case the inputs to fn will be indexable objects (tuples or arrays) of all channel values for a given sample, and fn should return a type-stable tuple value (for a multi-channel or single-channel result) or a number (for a single-channel result only). For example, the following would swap the left and right channels.

x = rand(10,2)
swapped = mapsignal(x,bychannel=false) do val
    val[2],val[1]
end

When bychannel=false the channels of each signal are not promoted:

Padding

Padding determines how samples past the end of shorter signals are reported. The value of padding is passd to pad. Its default value is determined by the value of fn. The default value for the four basic arithmetic operators is their identity (one for * and zero for +). These defaults are set on the basis of fn using default_pad(fn). A fallback implementation of default_pad returns zero.

To define a new default for a specific function, just create a new method of default_pad(fn)

myfun(x) = 2x + 3
SignalOperators.default_pad(::typeof(myfun)) = one

source

SignalOperators.mix — Function.

mix(xs...)

Sum all signals together, using mapsignal

source

SignalOperators.amplify — Function.

amplify(xs...)

Find the product, on a per-sample basis, for all signals xs using mapsignal.

source

SignalOperators.addchannel — Function.

addchannel(xs...)

Concatenate the channels of all signals into one signal, using mapsignal. This will result in a signal with sum(nchannels,xs) channels.

source

SignalOperators.channel — Function.

channel(x,n)

Select channel n of signal x, as a single-channel signal, using mapsignal.

source

SignalOperators.rampon — Function.

rampon(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds.

The function determines the shape of the ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.rampoff — Function.

rampoff(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the offset of a signal, smoothly transitioning from full amplitude to 0 amplitude over the course of len seconds.

The function determines the shape of the ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.ramp — Function.

ramp(x,[len=10ms],[fn=x -> sinpi(0.5x)])

Ramp the onset and offset of a signal, smoothly transitioning from 0 to full amplitude over the course of len seconds at the start and from full to 0 amplitude over the course of len seconds.

The function determines the shape of the ramp and should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source

SignalOperators.fadeto — Function.

fadeto(x,y,[len=10ms],[fn=x->sinpi(0.5x)])

Append x to y, with a smooth transition lasting len seconds fading from x to y (so the total length is duration(x) + duration(y) - len).

This fade is accomplished with a rampoff of x and a rampon for y. fn should be non-decreasing with a range of [0,1] over the domain [0,1]. It should map over the entire range: that is fn(0) == 0 and fn(1) == 1.

Both len and fn are optional arguments: either one or both can be specified, though len must occur before fn if present.

source